+++ CHANGELOG.md
@@ -46,3 +46,5 @@
 - add rem_vertices! returning a map from the new nodes to old ones
 - add dismantling
 - add cores and kcore
+- move from LightXML to EzXML
+- add missing read support for gexf